backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
Flask-Migrate==4.0.5